ionos-smtp/mcp-server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ionos-smtp-mcp"
version = "0.1.0"
description = "IONOS SMTP mail send MCP server for OpenClaw agents"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=3.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
rabbitmq = ["pika>=1.3"]
supabase = ["supabase>=2.0"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]

[project.scripts]
ionos-smtp-mcp = "ionos_smtp.server:main"
ionos-smtp-worker = "ionos_smtp.queue_worker:main"

[tool.hatch.build.targets.wheel]
packages = ["src/ionos_smtp"]
